Extract config class into module
